Delete Product fix

+++ client/src/Components/Admin/Products/DeleteProduct.tsx
@@ -1,25 +1,27 @@
-import { useEffect } from "react";
+import { useState, useEffect } from "react";
 import Swal from "sweetalert2";
 import withReactContent from "sweetalert2-react-content";
 
 const MySwal = withReactContent(Swal);
 
-const DeleteProduct = ({ onDelete }) => {
+const DeleteProduct = ({ onDelete, show }) => {
   useEffect(() => {
-    MySwal.fire({
-      title: "Are you sure?",
-      text: "You won't be able to revert this!",
-      icon: "warning",
-      showCancelButton: true,
-      confirmButtonColor: "#3085d6",
-      cancelButtonColor: "#d33",
-      confirmButtonText: "Yes, delete it!",
-    }).then((result) => {
-      if (result.isConfirmed) {
-        onDelete();
-      }
-    });
-  }, [onDelete]);
+    if (show) {
+      MySwal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!",
+      }).then((result) => {
+        if (result.isConfirmed) {
+          onDelete();
+        }
+      });
+    }
+  }, [onDelete, show]);
 
   return null;
 };
